Use undef_method to undefine all non-essential methods from the proxy class
Ideally, the proxy class should forward all method invocations to the underlying
object. But since many methods are defined in the Object class, they are
intercepted and invoked on the local proxy object. In 4f6b080 (Forward #to_s to
the remote object, 2010-11-06 18:00:11 +0100) I tried to work around that by
defining a custom #to_s method and simply forwarding that to the remote object.
Now I found a better way than to forward each method individually: simply
undef each method that we want to forward.